backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# ML & NLP
transformers==4.36.2
torch==2.1.2
nltk==3.8.1
spacy==3.7.2
scikit-learn==1.4.0
mlflow==2.9.2

# Data Processing
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# Twitter/X API
tweepy==4.14.0

# News API
newsapi-python==0.2.7
requests==2.31.0

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0
aiofiles==23.2.1

# Monitoring
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
faker==22.0.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
